+++ NanoXLSX.Shared/Enums/Styles/FontEnums.cs
@@ -27,7 +27,7 @@ namespace NanoXLSX.Shared.Enums.Styles
             none,
         }
         /// <summary>
-        /// Enum for the vertical alignment of the text from base line
+        /// Enum for the vertical alignment of the text from baseline
         /// </summary>
         public enum VerticalTextAlignValue
         {